+++ app/stores/UiStore.ts
@@ -1,6 +1,8 @@
 import { clamp } from "es-toolkit";
+import { t } from "i18next";
 import { action, computed, observable } from "mobx";
 import { flushSync } from "react-dom";
+import { toast } from "sonner";
 import { light as defaultTheme } from "@shared/styles/theme";
 import type { ProsemirrorData } from "@shared/types";
 import Storage from "@shared/utils/Storage";
@@ -486,13 +488,32 @@ class UiStore {
     this.debugSafeArea = !this.debugSafeArea;
   };
 
+  /**
+   * Display a toast for an export that is being prepared in the background,
+   * it is updated in place once the export completes or fails.
+   *
+   * @param fileOperationId The identifier of the export file operation.
+   */
   @action
-  registerExportToast = (
-    fileOperationId: string,
-    toastId: string,
-    timeoutId: ReturnType<typeof setTimeout>
-  ) => {
+  showExportToast = (fileOperationId: string) => {
+    const toastId = `export-${fileOperationId}`;
+
+    const timeoutId = setTimeout(() => {
+      toast.success(t("Export started"), {
+        id: toastId,
+        description: t("A link to your file will be sent through email soon"),
+        duration: 3000,
+      });
+      this.exportToasts.delete(fileOperationId);
+    }, 6000);
+
     this.exportToasts.set(fileOperationId, { toastId, timeoutId });
+
+    toast.loading(t("Export started"), {
+      id: toastId,
+      description: `${t("Preparing your download")}…`,
+      duration: Infinity,
+    });
   };
 
   @action
