+++ shared/i18n/locales/en_US/translation.json
@@ -409,9 +409,6 @@
   "our engineers have been notified": "our engineers have been notified",
   "Clear cache + reload": "Clear cache + reload",
   "Show detail": "Show detail",
-  "Export started": "Export started",
-  "A link to your file will be sent through email soon": "A link to your file will be sent through email soon",
-  "Preparing your download": "Preparing your download",
   "Exporting the collection may take some time.": "Exporting the collection may take some time.",
   "You will receive an email when it's complete.": "You will receive an email when it's complete.",
   "Include attachments": "Include attachments",
@@ -1518,6 +1515,9 @@
   "A confirmation code has been sent to your email address, please enter the code below to permanently destroy your account.": "A confirmation code has been sent to your email address, please enter the code below to permanently destroy your account.",
   "Are you sure? Deleting your account will destroy identifying data associated with your user and cannot be undone. You will be immediately logged out of {{appName}} and all your API tokens will be revoked.": "Are you sure? Deleting your account will destroy identifying data associated with your user and cannot be undone. You will be immediately logged out of {{appName}} and all your API tokens will be revoked.",
   "Delete my account": "Delete my account",
+  "Export started": "Export started",
+  "A link to your file will be sent through email soon": "A link to your file will be sent through email soon",
+  "Preparing your download": "Preparing your download",
   "Today": "Today",
   "Yesterday": "Yesterday",
   "Last week": "Last week",
